fix missing stuff (run-int-tests)
